+++ VoxelSharp.Abstractions/Renderer/IUpdatable.cs
@@ -2,5 +2,5 @@
 
 public interface IUpdatable
 {
-    public void Update(float deltaTime);
+    public void Update(double deltaTime);
 }
